+++ integration/setup/setup.go
@@ -63,7 +63,6 @@ func installResources(ctx context.Context, config Config) error {
 		"AppCatalog",
 		"AppCatalogEntry",
 		"Catalog",
-		"Chart",
 	}
 
 	{