+++ integration/setup/config.go
@@ -5,7 +5,6 @@ package setup
 
 import (
 	"github.com/giantswarm/apiextensions-application/api/v1alpha1"
-	"github.com/giantswarm/app/v6/pkg/crd"
 	"github.com/giantswarm/apptest"
 	"github.com/giantswarm/helmclient/v4/pkg/helmclient"
 	"github.com/giantswarm/k8sclient/v6/pkg/k8sclient"
@@ -20,7 +19,6 @@ import (
 
 type Config struct {
 	AppTest    apptest.Interface
-	CRDGetter  *crd.CRDGetter
 	HelmClient helmclient.Interface
 	K8s        *k8sclient.Setup
 	K8sClients k8sclient.Interface
@@ -56,15 +54,6 @@ func NewConfig() (Config, error) {
 		if err != nil {
 			return Config{}, microerror.Mask(err)
 		}
-	}
-
-	var crdGetter *crd.CRDGetter
-	{
-		c := crd.Config{
-			Logger: logger,
-		}
-
-		crdGetter, err = crd.NewCRDGetter(c)
 	}
 
 	var cpK8sClients *k8sclient.Clients
@@ -142,7 +131,6 @@ func NewConfig() (Config, error) {
 
 	c := Config{
 		AppTest:    appTest,
-		CRDGetter:  crdGetter,
 		HelmClient: helmClient,
 		K8s:        k8sSetup,
 		K8sClients: cpK8sClients,