+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+## [5.4.0] - 2021-12-17
+
 ### Changed
 
 - Update Helm to `v3.6.3`.
@@ -416,7 +418,8 @@ from Helm 2 to Helm 3.
 
 - Flattening operator release structure.
 
-[Unreleased]: https://github.com/giantswarm/app-operator/compare/v5.3.1...HEAD
+[Unreleased]: https://github.com/giantswarm/app-operator/compare/v5.4.0...HEAD
+[5.4.0]: https://github.com/giantswarm/app-operator/compare/v5.3.1...v5.4.0
 [5.3.1]: https://github.com/giantswarm/app-operator/compare/v5.3.0...v5.3.1
 [5.3.0]: https://github.com/giantswarm/app-operator/compare/v5.2.0...v5.3.0
 [5.2.0]: https://github.com/giantswarm/app-operator/compare/v5.1.1...v5.2.0